+++ apps/www/src/components/content/component-preview/index.astro
@@ -1,6 +1,6 @@
 ---
+import ExampleSource from "@/components/content/example-source.astro"
 import { cn } from "@/lib/utils"
-import ExampleSource from "../example-source.astro"
 import Preview from "./preview"
 import Tabs from "./tabs"
