+++ apps/www/src/layouts/root-layout.astro
@@ -33,7 +33,7 @@ import "@fontsource-variable/geist-mono"
 
     <ClientRouter />
 
-    <script src="@/lib/framework-script.ts"></script>
+    <script src="@/store/atoms/framework.ts" />
   </head>
   <body
     class="bg-background overscroll-none text-foreground font-sans antialiased"
